Initial fix for word check + word library
Added a word library

Added a normalize function for harakat, hmza, and other small letter differences for convenience

Fixed word check

Note: the game does NOT accept gibberish, so clearly the word library is working, but it still accepts so many obscure words it becomes too easy, it was hard to lose a life, need to review word library or scoring system.

Note: game startup is considerably slower, takes a few seconds for the website to respond on startup (likely due to loading the 50mb word library)

should look into a less resource intensive way of loading.